refactor: move observationCard to a separate file
